Add initial steps for a CLI client
There is a web-based frontend, but I could not get it to work with my
local testnet. Add a command line client as an alternative. With this I
am able to call CreateMultisig on the program deployed to my local
testnet.

The current code is just a quick initial program to get things running,
it does not do proper error handling yet, and it only connects to a
local testnet.