Simplify some code in transaction execution
Hello, if anyone's out there :wave:

I'm not sure if anyone uses this repo in production but it still seems
like a repo that people use to learn Anchor, judging by the 100 forks
and my own personal experience.

Case in point, I'm working on a timelock program & ended up re-using a
lot of the code from this repo. When I was looking at this part of it
though, I realized that you could accomplish this in a way that was more
simple and efficient. I've measured it and it removes 285 CUs in these
tests.

I think this provides pedagogical value even if this repo isn't being
used in prod.